Add files via upload
The data-set is a result of scraping the html code of (https://www.zomato.com/bangalore/great-food-no-bull) website, to collect data on all the finest restaurants in Bangalore .This data-set is created in the month of December, 2018. This data-set contain 11009 observations and 5 columns. Each observation shows the rating provided by a user for a particular restaurants in Bangalore.